% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SJC_boundary}
\alias{SJC_boundary}
\title{Polygon boundary dataset for San Juan County, Washington, USA}
\format{
`SJC_boundary`:
A data frame and sf object with 124 rows and 2 columns:
\describe{
  \item{island}{Name of the island that the polygon represents.}
  \item{geometry}{`sf` geometry.}
}
}
\usage{
SJC_boundary
}
\description{
An `sf` object containing the polygon boundaries of San Juan County,
Washington, USA.
}
\keyword{datasets}
