% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SJC_population}
\alias{SJC_population}
\title{Population dataset for San Juan County, Washington, USA}
\format{
`SJC_population`:
A data frame and sf object with 55,570 rows and 5 columns:
\describe{
  \item{island}{Name of the island that the pixel occurs on.}
  \item{tcc}{Auxiliary data. Tree canopy cover value at given pixel.}
  \item{elev}{Auxiliary data. Elevation value at given pixel.}
  \item{tnt}{Auxiliary data. Tree (2) and non-tree (1) layer.}
  \item{geometry}{`sf` geometry.}
}
}
\usage{
SJC_population
}
\description{
Population level data for each 90m pixel in San Juan County, Washington, USA.
For more information on the auxiliary data, see Wieczorek et al. (2023+).
}
\keyword{datasets}
