% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SJC_sample}
\alias{SJC_sample}
\title{Sample dataset for San Juan County, Washington, USA}
\format{
`SJC_sample`:
A data frame and sf object with 25 rows and 8 columns:
\describe{
  \item{plt_cn}{Name of the island that the pixel occurs on.}
  \item{biomass}{Average biomass measurement calculated from ground-collected
  FIA data. Units: short tons/acre.}
  \item{tcc}{Auxiliary data. Tree canopy cover value at given FIA plot.}
  \item{elev}{Auxiliary data. Elevation value at given FIA plot}
  \item{wc3cl}{Auxiliary data. Tree (2) and non-tree (1) value at a given 
  FIA plot.}
  \item{measurement_year}{Year in which the FIA plot was measured.}
  \item{island}{Name of the island that the FIA plot is on.}
  \item{geometry}{`sf` geometry.}
}
}
\usage{
SJC_sample
}
\description{
Survey data collected by the USDA Forest Service, Forest Inventory & Analysis
Program (FIA). This dataset contains plot-level data for biomass as collected 
and reported by the FIA. Further, auxiliary data has been matched with this
plot level data to match the population dataset.
}
\keyword{datasets}
